editoast: fga: move parts of client in `authorization_models`
